Modified AutonomousCommand PIDAutoDrive code to a) make the use of arguments to the drive code more clear and b) attempt to fix the inconsistency of distance moved during Autonomous.
This code compiles and runs on the robot but does not fix the problem.

The robot moves a portion of the requested distance (e.g. moves ~11 of
12 inches, ~4.5 feet of 6 feet).  It is consistent in how far it moves,
but does not move as far as requested.

The isFinished routine is checking to see if the target encoder value is
reached.
This target encoder value had been calculated using the absolute value
of the current encoder values before adding them.
Sean believes that the absolute value calculation was unnecessary (or
incorrect).  It was removed.
This change appears not to have made a difference, which I believes
makes sense.
Encoder 1 should be increasing (rolling "forwards") while Encoder 2
should be decreasing (rolling "backwards").  The current code adds
Encoder 1's value to the negative of Encoder 2's value.  This will have
the same effect as the previous absolute value code.  Further, the
previous absolute value code sholud work even if the Encoder 1 and
Encoder 2 wires are inadvertently crossed during wiring / rewiring of
the robot.
In either case, this change did not affect the behavior of the robot.

The initialization of the target encoder value also changed.  This one I
believe should have changed the behavior of the robot but appeared not
to - this is likely because the encoder values have been recently zeroed
in the initialization code so adding/subtracting a given value is the
same (values are zero).
The initialization had been adding the value of Encoder 2 without taking
the absolute value first (or multiplying by negative 1).
Before this change, only the first enqueued autonomous command was
executed.  After this change, both enqueued autonomous commands in my
test were executed.

Known bug - I tried a test of "drive 6 feet forward" followed by "drive
6 feet backwards".  These two commands executed, but the robot did not
stop after 6 feet on the return trip (it drove backwards but did not
stop, ever - it hit a table and started oscillating wildly so I
e-stopped it).  The code needs to be reviewed to make sure that a
negative speed (and positive distance) is the correct way to request a
move backwards.  The fix may be as simple as making the distance
negative as well.