New immutable configuration classes for complete configuration for an axis.
On a subsystem with a motor, there's different behavior for the motor, the encoder and the subsystem's exposed view of the hardware.

Obviously the human wants to program in human units like inches/feet/etc and the sensor and motor will give you back pulses and pulses/decisecond, but there's a lot of complexity in converting to reality on an actual motor:

The encoder, motor and drive shaft are often not on the same axis, each can be inverted, and each can have a gearing ratio relative to the others.  Abstracting away the nasty details so you can tell the arm to go up 10 inches or 1 ft/second or 20cm is necessary so you aren't constantly fighting bugs with conversion to pulses/decicm with arbitrary pulses per rotation, inversion and scaling.  It's a mess.  If something gets reversed somewhere it can break hundreds of places.

So... that leads me to the units library and a full description for each of the three logical axis.  I could probably have added a fourth, as sometimes you want to know about the drive shaft - like RPMs of the drive shaft, % of range rotated, but might also want to control how far the end of the arm swings, etc..

The next step is to switch all the values for speed/rate, length/distance, and time to the new units library.

Then taking all the behavior and translations for subsystem view of the device, the motor axis control units, and the sensor units/view of the device from these configuration classes.