STOP and OFF are now NEGATIVE.  Logically these are negative statements, not neutral.  They are unambiguous ways of saying something is NOT.
Added a function to return the most logical opposite value.  UP -> DOWN, this works well so that anytime you need to reverse the logic in a generic subsystem the value chosen can match the user's input.

Makes it easier if ON RIGHT UP doesn't become NEGATIVE instead of OFF, LEFT or DOWN.