Do not operate in auto.
This command is the default command for drive train.  When we're in auto, it runs any time the auto sequence isn't actively using the drive train.

We saw that it used the last buffered Joystick values and resulted in unwanted movement when auto wasn't driving.  Actually spinning back to the left (left side power reverse 9%) at idle.