Cleaned up Backward assumption for side starting position
Cleaned up assumptions in AutonomousCommandGroupGenerator that the L and
R staring positions will have the robot facing backwards.  We tried
that, but when the elevator sticks the top bar can block the shot.
Reverted to forward in most places, but not all.  Now controlled with a
variable.  Wouldn't need to be a variable if everything worked - more to
fix post-season.